requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
pillow
requests
aiohttp
python-multipart
typing
numpy
chromadb>=0.4.24
openai>=1.0.0
redis>=5.0.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
